changes splash screen logo
