+++ packages/manager/src/components/SplashScreen/keyframes.css
@@ -1,0 +1,69 @@
+/* Keyframes for the general loading icon animation */
+
+@keyframes moveupSmall {
+  0%, 60%, 100% {
+    transform: 
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(0);
+  }
+  25% {
+    transform:
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(0.5em);
+  }
+}
+
+@keyframes movedownSmall {
+  0%, 60%, 100% {
+    transform: 
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(0);
+  }
+  25% {
+    transform:
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(-0.5em);
+  }
+}
+
+@keyframes moveupLarge {
+  0%, 60%, 100% {
+    transform: 
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(0);
+  }
+  25% {
+    transform:
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(1.5em);
+  }
+}
+
+@keyframes movedownLarge {
+  0%, 60%, 100% {
+    transform: 
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(0);
+  }
+  25% {
+    transform:
+      rotateX(50deg)
+      rotateY(0deg)
+      rotateZ(45deg)
+      translateZ(-1.5em);
+  }
+}
